FileUtil: rework event handling and EOF logic to avoid truncation
Previiously, the incorrect part count estimation combined with
how the part count was used to determine how to handle EOF events
meant the fileUtil could declare completion prematurely, before the
last part of a file had been read and uploaded.

This reworks fileUtil to handle read and EOF events, and read calls,
all more or less the same via common helper which always drains the
bytes from the io stream, then determines if a segment is ready and
then if it is the last segment based on, and only on, the file EOF
having been seen and the stream fully read. This should avoid the
accidental truncated uploads previously seen when the EOF event
arrived after a read for a file with a size which meant it was
affected by the incorrect total part count calculation based on the
(lack of a) special-case for the size of the first part.